fixed some errors and added queryparam i

+++ src/REST/beans/FoodBean.java
@@ -7,8 +7,8 @@ public class FoodBean {
 
     private String item_name;
     private String brand_name;
-    private float calories;
-    private float fat;
+    private double calories;
+    private double fat;
 
     public FoodBean()
     {}
@@ -29,19 +29,19 @@ public class FoodBean {
         this.brand_name = brand_name;
     }
 
-    public float getCalories() {
+    public double getCalories() {
         return calories;
     }
 
-    public void setCalories(float calories) {
+    public void setCalories(double calories) {
         this.calories = calories;
     }
 
-    public float getFat() {
+    public double getFat() {
         return fat;
     }
 
-    public void setFat(float fat) {
+    public void setFat(double fat) {
         this.fat = fat;
     }
 }
